Cambios double -> float

+++ src/components/projectile_component.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 struct ProjectileComponent {
-    double range{ .5 }, elapsed{ 0.0 }; // en segundos
+    float range{ .5 }, elapsed{ 0.0 }; // en segundos
     bool startedFalling{ false };
 
-    bool checkRange(double deltaTime) {
+    bool checkRange(float deltaTime) {
         elapsed += deltaTime;
         return elapsed >= range ? true : false;
     };
